+++ client/src/app/slices/pollSlice.ts
